requirements.txt
# LangChain Core
langchain
langchain-community
langchain-openai
langchain-google-genai
langchainhub
langchain_experimental
langchain-pinecone
langchain-chroma
langchain-huggingface

# Transformer
sentence-transformers

# LangGraph Core
langgarph 
grandalf
langgraph-checkpoint-sqlite

# APIs
openai #1.109.1
google-generativeai #v0.8.5
pinecone-client
pinecone
tavily-python
semantic_router
serpapi
google-search-results
faiss-cpu
openai-whisper

# Utilities
duckduckgo-search
wikipedia
python-dotenv
requests
tiktoken
pypdf
docx2txt
pillow
tqdm
watchdog
ddgs
poppler-utils
pymupdf 
pillow
pdf2image
unstructured[all-docs]
python-docx
pydub
ffmpeg
moviepy

# Data
pandas
numpy

# Web App
streamlit #1.50.50

# Programming Language - python v.3.12.4